[chore] SupplementDetailViewModelOutput 주석 작성 및 SupplementDetailViewModel 파일 분리
